https://github.com/jpmorganchase/tessera/issues/447
Part of refactoring in preparation for Tessera issue #447.

Split console password reading from stream password reading.
When the JVM console isn't available, use an input stream (System.in at
runtime) to read the password from.

If not using the console, the password will be echoed, which may pose a
risk for some users. However, it is up to the user to run the
application in an environment that has a JVM console.