https://github.com/jpmorganchase/tessera/issues/447
Create parsers for each type of CLI option. Each parser is responsible
for it's own options and actions. This way, not everything is put into
one super CLI class.

Add ability to change, add or remove a password on a private key. This
option is only for file based private keys, and will stop other CLI
parsing from taking place, including but not limited to:
- new key generation
- configuration file loading and validating
- application startup

The rules and CLI options can be found in the attached issue.