Fix "No value present" error when trying to use admin CLI (#939)
* Cleanup mocks between tests & make tests more representative of reality
Changes made to improve tests:
* Explicitly define command properties as needed and cleanup before/after each test executes
* Do not define a subcommand without first defining a top-level command
* Add a help option to the MockCliAdapter.  The production code CliAdapters check for the presence of a help option and handle it accordingly.  The test fixture should do the same.

* Fix "No value present" error when trying to use admin CLI

Setting the CLI_TYPE_KEY system property when the "admin" CLI option was provided meant Admin CLI commands could not be executed.  The CliDelegate expects that a top-level command exists with the same type as that defined in the CLI_TYPE_KEY system property.  The AdminCliAdapter command is NOT top-level and so no valid adapter could be found, resulting in the "No value present" error.

By removing the setting of this property to ADMIN, the value of CLI_TYPE_KEY remains CONFIG.  The now CliDelegate tries to load the DefaultCliAdapter command as the top-level command, later adding the AdminCliAdapter command as a subcommand - this makes sense as it represents the usage of the admin commands, i.e. "tessera admin ...".

The existing test for the "admin" CLI option has been updated as this bug was not picked up by it. A specific mock for the AdminCliAdapter has been added and the test fixture now better represents real life - i.e. the mock config adapter is the top-level command, the mock admin adapter is a subcommand.