Add extra logging around PartyInfo
Adds extra logging around various stages of PartyInfo so there is more
details when debugging issues. Most of the extra logs are DEBUG since
they are not needed during normal operation, but there are a few INFO
level messages that give the user a view that the application is
ticking along.