Remove inner resend loop
There are two loops that occurs when perform a resend. The first will
try 20 times, with a backoff period set by the poller rate (user
configurable). The intention is that a resend will occur eventually if
one of the nodes was offline, but not to try indefinitely.

The second loop occurs 5 times *for each* of the 20 iterations
mentioned prior. This was intended to catch ephemeral issues, but
causes more harm than good. Any ephemeral issue will be caught during
the 20-loop, and even so, after 5 failed attempts it would be pretty
clear the issue is persistent.

This commit removes the 5 attempts per loop, greatly reducing the
number of retries made from 100 to 20, and simplifying the flow.